refactor: data models add: offline data source model
