+++ lib/feature/exam/data/apis/DTO/subject_dto.dart
@@ -1,4 +1,4 @@
-import 'package:quizz_app/feature/exam/data/data_sources/offline_data_source/subject_cached_model.dart';
+import 'package:quizz_app/feature/exam/data/data_sources/offline_data_source/models/subjects_model/subject_cached_model.dart';
 import 'package:quizz_app/feature/exam/data/models/subject_response.dart';
 import '../../../domain/entities/exam_entity.dart';
 import '../../../domain/entities/subject_item_entity.dart';
