+++ lib/feature/exam/data/core/caching_data.dart
@@ -2,7 +2,7 @@ import 'package:hive/hive.dart';
 
 import '../../../../core/resources/app_constant.dart';
 import '../../domain/entities/subject_item_entity.dart';
-import '../data_sources/offline_data_source/subject_cached_model.dart';
+import '../data_sources/offline_data_source/models/subjects_model/subject_cached_model.dart';
 
 abstract class CachingData{
   static void cachedSubject(List<SubjectItemEntity> subjects) {
